DynamicTablesPkg: Add RISC-V specific table definitions temporarily
For RISC-V, there are new structures defined in MADT and a new static
table RHCT is defined. However, though ECRs are approved the next
version of the spec with these content is not released yet. So, add
temporary definitions till ACPI spec released and a new header file is
added in EDK2 source.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>