DynamicTablesPkg: AcpiFadtLib: Move ArmBootArch to ARM specific file
ArmBootArch field in FADT is specific to ARM. To leverage most of the
code in FADT generator for other architectures, this needs to be moved
into ARM specific file. There may be other fields in future (ex:
IA_PC_BOOT_ARCH) that need to be created in similar way. So, instead of
adding the function only to create BootArch, make this function generic
to add all arch specific functions into FADT.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>