DynamicTablesPkg: Rename ArmNameSpaceObjects to ArchNameSpaceObjects
Currently, ArmNameSpaceObjects defines information which is ARM specific
as well as common across architectures. There may be few more
architecture specific information for each architecture. Creating
multiple Namespace object types for each architecture and one for
common seems to be overkill and doesn't help much is memory
requirements. So, just rename the current ArmNameSpaceObjects to
ArchNameSpaceObjects so that any architecture can make use of these
definitions.

Since the change affects multiple files and platforms, the patch is
quite big. But if the approach is fine, this can be split into multiple
patches according to driver/module/platform.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>