MdePkg/DxeRiscvMpxyLib: Enable per hart shared memory
Since at OS run time, the MPXY can be called on any CPU, each of those
harts which send message should setup the shared memory. Currently, it
is assumed that it is used on only one cpu which is fine during EDK2
boot. But once OS starts, the SMP will be enabled. Also, at the OS
runtime, the shared memory will be virtual but the SBI call expects
physical address. To acommodate these nuiances, update the prototype of
SetShmem() and use virtual or physical address appropriately.

Since all cpus are not available at EDK2 boot time, it is not possible
to use different shared memory for each hart. Since EFI calls are
guranteed to be single threaded by OS level global lock, it may be safer
to assume same shared memory can be set on every hart calling
send_message.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>